Rename instead of scool to school
