+++ pom.xml
@@ -9,7 +9,7 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>ru.hogwarts</groupId>
-    <artifactId>AdditionalCourseScool</artifactId>
+    <artifactId>AdditionalCourseSchool</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>AdditionalCourseScool</name>
     <description>AdditionalCourseScool</description>
