+++ src/main/java/ru/hogwarts/additionalcourseschool/controllers/StudentController.java
@@ -1,4 +1,4 @@
-package ru.hogwarts.additionalcoursescool.controllers;
+package ru.hogwarts.additionalcourseschool.controllers;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -11,309 +11,309 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import ru.hogwarts.additionalcoursescool.dto.FacultyDTO;
-import ru.hogwarts.additionalcoursescool.model.Faculty;
-import ru.hogwarts.additionalcoursescool.services.FacultyService;
+import ru.hogwarts.additionalcourseschool.dto.StudentDTO;
+import ru.hogwarts.additionalcourseschool.model.Student;
+import ru.hogwarts.additionalcourseschool.services.StudentService;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/faculty")
-@Tag(name = "Faculties", description = "CRUD-operations to work with the faculties")
-public class FacultyController {
-    private final FacultyService facultyService;
+@RequestMapping("/student")
+@Tag(name = "Students", description = "CRUD-operations to work with the students")
+public class StudentController {
+private final StudentService studentService;
 
-    public FacultyController(FacultyService facultyService) {
-        this.facultyService = facultyService;
+    public StudentController(StudentService studentService) {
+        this.studentService = studentService;
     }
 
     @PostMapping
     @Operation(
-            summary = "Create new faculty",
-            description = "Create new faculty with its number"
+            summary = "Create new student",
+            description = "Create new student with its number"
     )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculty was successfully created",
+                    description = "Student was successfully created",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             )
     })
-    public ResponseEntity<FacultyDTO> createFaculty(@RequestBody FacultyDTO facultyDTO) {
-        FacultyDTO createdFacultyDTO = facultyService.createFaculty(facultyDTO);
-        return ResponseEntity.ok(createdFacultyDTO);
+    public ResponseEntity<StudentDTO> createStudent(@RequestBody StudentDTO studentDTO) {
+        StudentDTO createdStudentDTO = studentService.createStudent(studentDTO);
+        return ResponseEntity.ok(createdStudentDTO);
     }
 
-    @GetMapping("/{facultyId}")
+    @GetMapping("/{studentId}")
     @Operation(
-            summary = "Find faculty by its number",
-            description = "Search by faculty number"
+            summary = "Find student by its number",
+            description = "Search by student number"
     )
     @Parameters(value = {
-            @Parameter(name = "facultyId", example = "1")
+            @Parameter(name = "studentId", example = "1")
     })
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculty was successfully found",
+                    description = "Student was successfully found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             ),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Faculty not found",
+                    description = "Student not found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             )
     })
-    public ResponseEntity<FacultyDTO> getFaculty(@PathVariable Long facultyId) {
-        FacultyDTO facultyDTO = facultyService.findFacultyById(facultyId);
-        if (facultyDTO == null) {
+    public ResponseEntity<StudentDTO> getStudent(@PathVariable Long studentId) {
+        StudentDTO studentDTO = studentService.findStudentById(studentId);
+        if (studentDTO == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(facultyDTO);
+        return ResponseEntity.ok(studentDTO);
     }
 
-    @GetMapping("/findByColor")
+    @GetMapping("/findByAge")
     @Operation(
-            summary = "Find faculty list by color",
-            description = "Show faculty list by color"
+            summary = "Find student list by age",
+            description = "Show student list by age"
     )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculty list was successfully found",
+                    description = "Student list was successfully found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             ),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Faculty list was not found",
+                    description = "Student list was not found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             )
     })
-    public ResponseEntity<List<FacultyDTO>> getFacultyByColor(@RequestParam String color) {
-        List<FacultyDTO> facultyDTOS = facultyService.findFacultyByColor(color);
-        if (facultyDTOS == null) {
+    public ResponseEntity<List<StudentDTO>> getStudentByAge(@RequestParam int age) {
+        List<StudentDTO> studentDTOS = studentService.findStudentByAge(age);
+        if (studentDTOS == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(facultyDTOS);
+        return ResponseEntity.ok(studentDTOS);
     }
 
-    @GetMapping("/findFacultyByName")
+    @GetMapping("/findByAgeBetween")
     @Operation(
-            summary = "Find faculty by name",
-            description = "Show faculty by name"
+            summary = "Find student list by age between",
+            description = "Show student list by age between"
     )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculty was successfully found",
+                    description = "Student list was successfully found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             ),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Faculty was not found",
+                    description = "Student list was not found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             )
     })
-    public ResponseEntity<FacultyDTO> getFacultyByName(@RequestParam String name) {
-        FacultyDTO facultyDTO = facultyService.findFacultyByNameIgnoreCase(name);
-        if (facultyDTO == null) {
+    public ResponseEntity<List<StudentDTO>> getStudentByAgeBetween(@RequestParam int min, @RequestParam int max) {
+        List<StudentDTO> studentDTOS = studentService.findStudentByAgeBetween(min, max);
+        if (studentDTOS == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(facultyDTO);
+        return ResponseEntity.ok(studentDTOS);
     }
 
-    @GetMapping("/findByStudentId")
+    @GetMapping("/getStudentsByFacultyId")
     @Operation(
-            summary = "Find faculty by student id",
-            description = "Show faculty by student id"
+            summary = "Find student by faculty id",
+            description = "Search by faculty id"
     )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculty was successfully found",
+                    description = "Student was successfully found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             ),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Faculty was not found",
+                    description = "Student not found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             )
     })
-    public ResponseEntity<FacultyDTO> getFacultyByStudentId(@RequestParam Long studentId) {
-        FacultyDTO facultyDTO = facultyService.findFacultyByStudentId(studentId);
-        if (facultyDTO == null) {
-            return ResponseEntity.notFound().build();
+    ResponseEntity<List<StudentDTO>> getStudentsByFacultyId(@RequestParam Long facultyId) {
+        List<StudentDTO> studentDTOS = studentService.findStudentByFacultyId(facultyId);
+        if (studentDTOS == null) {
+            ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(facultyDTO);
+        return ResponseEntity.ok(studentDTOS);
     }
 
     @GetMapping
     @Operation(
-            summary = "Find all faculties",
-            description = "Show all faculties"
+            summary = "Find all students",
+            description = "Show all students"
     )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculties were successfully found",
+                    description = "Students were successfully found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             ),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Faculties were not found",
+                    description = "Students were not found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             )
     })
-    public ResponseEntity<List<FacultyDTO>> getAllFaculties() {
-        List<FacultyDTO> facultyDTOS = facultyService.findAllFaculties();
-        if (facultyDTOS == null) {
+    public ResponseEntity<List<StudentDTO>> getAllStudents() {
+        List<StudentDTO> studentDTOS = studentService.findAllStudents();
+        if (studentDTOS == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(facultyDTOS);
+        return ResponseEntity.ok(studentDTOS);
     }
 
     @PutMapping
     @Operation(
-            summary = "Update faculty by its number",
-            description = "Search by faculty number to update it"
+            summary = "Update student by its number",
+            description = "Search by student number to update it"
     )
     @Parameters(value = {
-            @Parameter(name = "facultyId", example = "1")
+            @Parameter(name = "studentId", example = "1")
     })
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculty was successfully updated",
+                    description = "Student was successfully updated",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             ),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Faculty not found",
+                    description = "Student not found",
                     content = {
                             @Content(
                                     mediaType = "application/json",
                                     array = @ArraySchema(schema =
-                                    @Schema(implementation = Faculty.class))
+                                    @Schema(implementation = Student.class))
                             )
                     }
             )
     })
-    public ResponseEntity<FacultyDTO> updateFaculty(@RequestBody FacultyDTO facultyDTO) {
-        FacultyDTO updatedFacultyDTO = facultyService.updateFaculty(facultyDTO);
-        if (updatedFacultyDTO == null) {
+    public ResponseEntity<StudentDTO> updateStudent(@RequestBody StudentDTO studentDTO) {
+        StudentDTO updatedStudentDTO = studentService.updateStudent(studentDTO);
+        if (updatedStudentDTO == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(updatedFacultyDTO);
+        return ResponseEntity.ok(updatedStudentDTO);
     }
 
-    @DeleteMapping("/{facultyId}")
+    @DeleteMapping("/{studentId}")
     @Operation(
-            summary = "Delete faculty by its number",
-            description = "Search by faculty number to delete it"
+            summary = "Delete student by its number",
+            description = "Search by student number to delete it"
     )
     @Parameters(value = {
-            @Parameter(name = "facultyId", example = "1")
+            @Parameter(name = "studentId", example = "1")
     })
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculty was successfully deleted"
+                    description = "Student was successfully deleted"
             ),
             @ApiResponse(
                     responseCode = "404",
-                    description = "Faculty not found"
+                    description = "Student not found"
             )
     })
-    public ResponseEntity<Void> deleteFacultyById(@PathVariable Long facultyId) {
-        facultyService.deleteFaculty(facultyId);
+    public ResponseEntity<Void> deleteStudentById(@PathVariable Long studentId) {
+        studentService.deleteStudent(studentId);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping
     @Operation(
-            summary = "Delete all faculties",
-            description = "Delete all faculties"
+            summary = "Delete all students",
+            description = "Delete all students"
     )
     @ApiResponses(value = {
             @ApiResponse(
                     responseCode = "200",
-                    description = "Faculties were successfully deleted"
+                    description = "Students were successfully deleted"
             )
     })
-    public ResponseEntity<Void> deleteAllFaculties() {
-        facultyService.deleteAllFaculties();
+    public ResponseEntity<Void> deleteAllStudents() {
+        studentService.deleteAllStudents();
         return ResponseEntity.ok().build();
     }
 }
